% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_plot_utils.R
\name{get_variable_importance}
\alias{get_variable_importance}
\title{Gets variable importance graph for Random Forest model (ranger package)}
\usage{
get_variable_importance(
  model_data,
  target_col,
  predictors,
  predictor_labels = NULL,
  classification = TRUE,
  probability = TRUE
)
}
\arguments{
\item{model_data}{data frame containing the data to build the model on. It is
assumed that the specified target and predictors are all present in this df.}

\item{target_col}{name of the target column}

\item{predictors}{list of predictor columns to consider. It is presumed that
all of these columns are present in the model_data df.}

\item{predictor_labels}{optional list of user-friendly names for each
predictor variable. If supplied, then there should be a value for each of
the predictor variables.}

\item{classification}{TRUE/FALSE flag to indicate if the Random Forest is a
classification model (as opposed to regression)}

\item{probability}{TRUE/FALSE flag to indicate if the Random Forest model
should generate probabilities}
}
\description{
\code{get_variable_importance} Generates a plot showing the estimated
imporance of predictor variables for a Randome Forest model using the ranger
package
}
\details{
Use this for a quick assessment of which variables are most important when
building a model
}
