% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsec_postproc_utils.R
\name{rr_add_subseg_failure_class_sums}
\alias{rr_add_subseg_failure_class_sums}
\title{Gets the sum of deficit scores in a segment for a set of data_codes}
\usage{
rr_add_subseg_failure_class_sums(sub_segs, deficit_data, failure_type_lkps)
}
\arguments{
\item{sub_segs}{sub-sectioning output data frame. This data
frame should contain all distresses or data relating to capacity deficits,
and should contain the following columns:
\enumerate{
\item section_id: containing the unique ID for sections on which deficit
data is found
\item loc_from: start location for each deficit observation
\item loc_to: end location for each deficit observation
\item lane: lane code
\item deficit: normalised deficit score for each observation. It is
recommended that a scale of 0 to 10 be used, with higher values indicating
greater deficit (distress or capacity deficit)
\item data_code: short code identifying each data type in the set
}}

\item{deficit_data}{data frame with deficit data for the network. This data
frame should contain all distresses or data relating to capacity deficits,
and should contain the following columns:
\enumerate{
\item section_id: containing the unique ID for sections on which deficit
data is found
\item loc_from: start location for each deficit observation
\item loc_to: end location for each deficit observation
\item lane: lane code
\item deficit: normalised deficit score for each observation. It is
recommended that a scale of 0 to 10 be used, with higher values indicating
greater deficit (distress or capacity deficit)
\item data_code: short code identifying each data type in the set
}}

\item{failure_type_lkps}{data frame containing a lookup set in which each
data_code found in the deficit data is associated with a more general
failure type. If there are any data codes in the deficit data that are not
handled in the failure type lookups, then an error is thrown. This data
frame should have columns:
\enumerate{#'
\item data_code: short code identifying each data type in the set
\item failure_type: short code representing the more general failure type
associated with each data code
}}
}
\description{
\code{rr_add_subseg_failure_class_sums} gets the deficit sums for specified
failure classes associated with each deficit data type, and adds a column for
each failure class to the sub_segment data set.
}
\details{
This function allows you to get an added column that has, for example, the
sum of deficit scores for all data codes associated with a certain failure
type such as 'deformation' or 'shear failure'
}
