% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engineering.R
\name{rr_classify_defects_absol}
\alias{rr_classify_defects_absol}
\title{Assigns a percentile based 0 to 4 score based on defect values (absolute)}
\usage{
rr_classify_defects_absol(df, col_name, c1, c2, c3, as_numeric = TRUE)
}
\arguments{
\item{df}{data frame holding the data}

\item{col_name}{name of the column that holds the defect data values}

\item{c1}{upper threshold for class 1 (zeros excluded)}

\item{c2}{upper threshold for class 2}

\item{c3}{upper threshold for class 3. Values greater than this threshold
are assigned a class 4 automatically}

\item{as_numeric}{If TRUE (default) then the values returned are numeric
(0,1,2,3,4) instead of Factor (be careful to convert Factor to numeric!)}
}
\value{
vector containing class values for each defect value
}
\description{
\code{rr_classify_defects_absol} can be used to get a classification of
0 to 4 on a set of defect values that are assumed to be zero when there is no
defect, and a high number (can be any number) when there is severe/extensive
defects.
}
\details{
The assignment of classes is made such that the assigned classes are:
\enumerate{
\item 0: Assigned when defect values are zero
\item 1: Assigned to values lower parameter c1 but NOT zero
\item 2: Assigned to defect values that lie between the c1 and c2
\item 3: Assigned to defect values that lie between the c2 and c3
\item 4: Assigned to defect values that are above c3
}

This function can be used to assign a severity scale based on absolute
threshold values. Defect values that are zero are assigned a zero class, and
classes 1 to 4 are assigned based on the thresholds provided as parameters
c1, c2 and c3

Use this function when the distress values are reliably mapped to
absolute values such as percent of segment or square metre. Otherwise, you
can use a relative assignement based on percentiles:
}
\seealso{
\code{link{rr_classify_defects_relative}}
}
