% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engineering.R
\name{rr_classify_defects_relative}
\alias{rr_classify_defects_relative}
\title{Assigns a percentile based 0 to 4 score based on defect values (relative)}
\usage{
rr_classify_defects_relative(
  df,
  col_name,
  perc_c1 = 0.25,
  perc_c2 = 0.5,
  perc_c3 = 0.75,
  as_numeric = TRUE
)
}
\arguments{
\item{df}{data frame holding the data}

\item{col_name}{name of the column that holds the defect data values}

\item{perc_c1}{percentile that is the upper cut-off for class 1}

\item{perc_c2}{percentile that is the upper cut-off for class 2}

\item{perc_c3}{percentile that is the upper cut-off for class 3. Values
greater than this percentile are assigned a class 4 automatically}

\item{as_numeric}{If TRUE (default) then the values returned are numeric
(0,1,2,3,4) instead of Factor (be careful to convert Factor to numeric)}
}
\value{
vector containing class values for each defect value
}
\description{
\code{rr_classify_defects_relative} can be used to get a classification of
0 to 4 on a set of defect values that are assumed to be zero when there is no
defect, and a high number (can be any number) when there is severe/extensive
defects.
}
\details{
The assignment of classes is made such that the assigned classes are:
\enumerate{
\item 0: Assigned when defect values are zero
\item 1: Assigned to values lower then the 25th percentile of values that
are NOT zero
\item 2: Assigned to defect values that lie between the 25th to 50th
percentiles of values that are NOT zero
\item 3: Assigned to defect values that lie between the 50th to 75th
percentiles of values that are NOT zero
\item 4: Assigned to defect values that are above the 75th percentile of
values that are NOT zero
}

This function can be used to assign a relative severity scale that ensures
that defect values that are zero are assigned a zero class, while the values
that are above a certain percentile get a rating of severe = 4.

This function is useful when the distress values are not reliably mapped to
absolute values such as percent of segment or square metre. It can handle
zero inflated situations that are highly skewed.

Note that the percentiles noted above (25th, 50th, 75th) are defaults and
can be overwritten in the parameter specifications.
}
