% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deficit_set_utils.R
\name{rr_Clean_deficit_set}
\alias{rr_Clean_deficit_set}
\title{Cleans a deficit input set for sub-sectioning and checks for problems}
\usage{
rr_Clean_deficit_set(deficit_set, min_obs_length = 1, max_obs_length = 30)
}
\arguments{
\item{deficit_set}{data frame with deficit data for the segment. This data frame
should contain all distresses or data relating to capacity deficits, and
should contain the following columns:
\enumerate{
\item section_id: containing the unique ID for sections
\item loc_from: start location for each deficit observation
\item loc_to: end location for each deficit observation
\item deficit: normalised deficit score for each observation. It is
recommended that a scale of 0 to 10 be used, with higher values indicating
greater deficit (distress or capacity deficit)
\item data_code: short code identifying each data type in the set
}}

\item{min_obs_length}{minimum length for an observation. It is strongly
recommended that this value be set at 1m. Values of less than one will be
corrected in any event by the sub-sectioning algorithm}

\item{max_obs_length}{maximum length for a deficit observation. This value will
ensure that a single observation that has a low deficit but a very long value
does not dominate the selection of a sub-section. If a value is longer than this
specified maximum length, the from and to locations will be modified so that the
observation has the same middle but length equal to the allowed maximum}
}
\description{
\code{rr_Clean_deficit_set} Checks a prepared input set for sub-sectioning
and makes corrections to the start and end locations for each observation.
}
\details{
Currently, the primary objective of this function is to ensure that there are
no very long observations that will dominate sub-section identification. For
example, a flushing area may have a low deficit but if it is 70m long the
deficit score (which multiplies the deficit by the length) will push this
observation to the top because of its length. If you want to prevent these
types of situations, then use a shorter maximum length such as 30m.
}
