% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{rr_get_boolean}
\alias{rr_get_boolean}
\title{Gets a boolean value from a general value}
\usage{
rr_get_boolean(value)
}
\arguments{
\item{value}{value (character or number) to evaluate}
}
\value{
TRUE or FALSE
}
\description{
\code{tt_get_boolean} returns TRUE or FALSE based on the following
conditions:
\enumerate{
\item if value is NULL or NA returns FALSE
\item if value as numeric is 1 returns TRUE
\item if value as numeric is 0 returns FALSE
\item if lowercase of value is any of 'true', 't', 'yes' or 'y':
return TRUE
\item if lowercase of value is any of 'false', 'f', 'no' or 'n':
return FALSE
\item if length of value as character after trimming is 0, return FALSE
\item if none of the above, return TRUE
}
}
