% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_sectioning_v1.R
\name{rr_get_deficit_score}
\alias{rr_get_deficit_score}
\title{Gets the Deficit Score for a specific sub-section}
\usage{
rr_get_deficit_score(df, loc_from, loc_to, benefit_scaler, method = "npv")
}
\arguments{
\item{df}{data frame with deficit data for the section}

\item{loc_from}{start location for the sub-section}

\item{loc_to}{end location for the sub-section}

\item{benefit_scaler}{maximum location for the segment}

\item{method}{method to use for calculating optimal sub-sectioning. Valid
values are 'npv' or 'bcr'. The 'npv' method is recommended. 'bcr' tends to
give very short sub-sections}
}
\value{
score for the sub-section limits
}
\description{
\code{rr_get_deficit_score} Allows you to manually calculate the Deficit
Score for a specific sub-section.
}
