% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_visual_surv.R
\name{rr_get_distress_pct}
\alias{rr_get_distress_pct}
\title{Converts a distress quantity to a percentage}
\usage{
rr_get_distress_pct(
  df,
  distress_column,
  insp_from_col,
  insp_to_col,
  width_col = "none",
  area_based = FALSE
)
}
\arguments{
\item{df}{data frame with the distress values and inspection start and end
locations}

\item{distress_column}{name of the column containing the distress to convert}

\item{insp_from_col}{name of the column containing the inspection START
locations}

\item{insp_to_col}{name of the column containing the inspection END
locations}

\item{width_col}{name of the column containing the width of the inspection
area. Only used if parameter \code{area_based} is TRUE.}

\item{area_based}{TRUE/FALSE value indicating whether the percentage
conversion is area or length based. If area based, then the inspection length
is converted to a area using the specified width column}
}
\value{
vector of percentage values, with percentages as 0 to 100+ values.
}
\description{
\code{rr_get_distress_pct} converts a quantity (either length or area) of
observed distress to a percentage value based on the observed quantity and
the inspection length.
}
\details{
For example, if a 25 m length of cracking was observed in an inspection
length ranging from location 50 to 150, then that would equate to 25m distress
for an inspection length of 100 m, thus the percentage is 25\%.

This function does the conversion for all values in a data frame and returns
a vector with the percentage values. It only does one distress at a time.
}
