% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{rr_get_lookup_value}
\alias{rr_get_lookup_value}
\title{Does secure lookup on a lookup set}
\usage{
rr_get_lookup_value(
  lookups,
  lookup_set_key,
  key_value,
  key_column = "key",
  lookup_column = "value"
)
}
\arguments{
\item{lookups}{List object containing lookup sets read with method
\code{tt_get_lookups}}

\item{lookup_set_key}{key that identifies the required lookup set. If this
key is not found in the names of parameter lookups, an error is thrown}

\item{key_value}{key that identifies the row on which the value should be
retrieved. If the key does not exist, an error is thrown}

\item{key_column}{column that contains the key on which to look up on. If
the column does not exist, an informative error is thrown}

\item{lookup_column}{column on which to look up they value on. If the column
does not exist, an informative error is thrown}
}
\description{
\code{tt_get_lookup_value} Finds the value in a required column of a
specified lookup table. If either the lookup table key or the lookup column
does not exist an informative error is thrown.
}
