% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_visual_surv.R
\name{rr_get_pothole_pct}
\alias{rr_get_pothole_pct}
\title{Converts a column with pothole counts to a percentage of inspection area}
\usage{
rr_get_pothole_pct(
  df,
  pothole_column,
  pothole_area,
  insp_from_col,
  insp_to_col,
  width_col
)
}
\arguments{
\item{df}{data frame with the pothole counts and inspection start and end
locations}

\item{pothole_column}{name of the column containing the pothole counts}

\item{pothole_area}{square metre area assumed for each pothole}

\item{insp_from_col}{name of the column containing the inspection START
locations}

\item{insp_to_col}{name of the column containing the inspection END
locations}

\item{width_col}{name of the column containing the width of the inspection
area}
}
\value{
vector of percentage area values, with percentages as 0 to 100+ values.
}
\description{
\code{rr_get_pothole_pct} converts a pothole count to a percentage value based on the observed quantity and
the inspection area. The area each pothole is assumed to represent can be
specified
}
