% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_utils.R
\name{rr_get_seg_data_ovlp}
\alias{rr_get_seg_data_ovlp}
\title{Gets the row with maximum overlap over a specified minimum overlap}
\usage{
rr_get_seg_data_ovlp(seg, df, min_overlap = 0.5)
}
\arguments{
\item{seg}{data frame containing information for the specified
segment. Required columns are:
\enumerate{
\item section_id: ID for the section on which the treatment length is
\item loc_from: Start/From location for the treatment length
\item loc_to: End/To location for the treatment length
\item lane: Lane code for the treatment length
}}

\item{df}{data frame with data to filter for the treatment length. This
data frame should contain the following columns:
\enumerate{
\item section_id: containing the unique ID for sections on which deficit
data is found
\item loc_from: start location for each deficit observation
\item loc_to: end location for each deficit observation
\item lane: data lane code, or 'all' if the data applies to all lanes
}}

\item{min_overlap}{minimum overlap percent using treatment length as base}
}
\value{
The row with the maximum overlap, IF the overlap is more than the
specified minimum fraction of the length of the treatment. If there are no
overlapping rows, or if the maximum overlap is less than the specified
minimum overlap, then NULL is returned
}
\description{
\code{rr_get_seg_data_ovlp} Filters data and returns only those rows where
(a) the section Id matches the section ID of the specified treatment length;
(b) the data locations has at least a certain overlap with treatment length
locations
(c) the data lane code is 'all' or else matches the lane code of the
specified treatment length (if the treatment length has lane code 'all', then
lane filtering is omitted)
}
