% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsec_utils.R
\name{rr_get_subsecs_on_sections}
\alias{rr_get_subsecs_on_sections}
\title{Gets optimal subsections on a set of Sections}
\usage{
rr_get_subsecs_on_sections(
  sections,
  deficit_data,
  min_length,
  max_length,
  score_threshold,
  benefit_scaler,
  method = "npv",
  max_subsecs = 25,
  ends_join_limit = 50
)
}
\arguments{
\item{sections}{data frame containing a list of all sections on which
sub-sectioning should be done. This dataframe should contain the following
columns:
\enumerate{
\item section_id: containing the unique ID for sections
\item section_name: name of the section
\item loc_from: start location for each section
\item loc_to: end location for each section
}}

\item{deficit_data}{data frame with deficit data for the network This data
frame should contain all distresses or data relating to capacity deficits,
and should contain the following columns:
\enumerate{
\item section_id: containing the unique ID for sections on which deficit
data is found
\item loc_from: start location for each deficit observation
\item loc_to: end location for each deficit observation
\item deficit: normalised deficit score for each observation. It is
recommended that a scale of 0 to 10 be used, with higher values indicating
greater deficit (distress or capacity deficit)
\item data_code: short code identifying each data type in the set
}}

\item{min_length}{minimum sub-section length allowed}

\item{max_length}{maximum sub-section length allowed}

\item{score_threshold}{lower threshold below which sub-sections should be
ignored}

\item{benefit_scaler}{scaling factor for deficits. This factor should be
manually calibrated using the plotting function}

\item{method}{method to use for calculating optimal sub-sectioning. Valid
values are 'npv' or 'bcr'. The 'npv' method is recommended. 'bcr' tends to
give very short sub-sections}

\item{max_subsecs}{maximum number of sub-sections to allow within each
section (default is 25)}

\item{ends_join_limit}{length within which any sub-sections close to the
start or end of a section should be joined to the start or end. Default
value is 50m. Pass in a value of -1 if you do NOT want to join sub-sections
to the starts of ends.}
}
\description{
\code{rr_get_subsecs_on_sections} Generates a set of optimal sub-sections
for each section within a larger set. For example, you can use this function
to generate sub-sections for an entire network when the set of sections
provided as the first parameter represent all sections on your network.
}
\details{
Note: This method does NOT provide a full set of segments representing the
entire network as defined by the list of Sections you proivide. It only
provides a list of optimal sub-sections for the network based on the
parameters \code{score_threshold} and \code{benefit_scaler}.

Also note that this method, unlike \code{\link{rr_get_fwp_subsecs}} is NOT
lane specific.
}
