% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsec_postproc_utils.R
\name{rr_get_subseg_defect_sums}
\alias{rr_get_subseg_defect_sums}
\title{Gets the sum of deficit scores in a segment for a set of data_codes}
\usage{
rr_get_subseg_defect_sums(sub_seg, deficit_data, data_codes_to_include)
}
\arguments{
\item{sub_seg}{row from sub-sectioning output data frame. This data
frame should contain all distresses or data relating to capacity deficits,
and should contain the following columns:
\enumerate{
\item section_id: containing the unique ID for sections on which deficit
data is found
\item loc_from: start location for each deficit observation
\item loc_to: end location for each deficit observation
\item lane: lane code
\item deficit: normalised deficit score for each observation. It is
recommended that a scale of 0 to 10 be used, with higher values indicating
greater deficit (distress or capacity deficit)
\item data_code: short code identifying each data type in the set
}}

\item{deficit_data}{data frame with deficit data for the network. This data
frame should contain all distresses or data relating to capacity deficits,
and should contain the following columns:
\enumerate{
\item section_id: containing the unique ID for sections on which deficit
data is found
\item loc_from: start location for each deficit observation
\item loc_to: end location for each deficit observation
\item lane: lane code
\item deficit: normalised deficit score for each observation. It is
recommended that a scale of 0 to 10 be used, with higher values indicating
greater deficit (distress or capacity deficit)
\item data_code: short code identifying each data type in the set
}}

\item{data_codes_to_include}{vector containing the data codes to include in
the sum.}
}
\description{
\code{rr_get_subseg_defect_sums} gets the sum of the deficit scores within a
segment for all data codes within a set of values passed in parameter
\code{data_codes_to_include}.
}
\details{
Use this function to get the sum of deficit scores for a specific class of
defect, e.g. those data_codes mapping to deformation or shear etc. Here,
deficit score means the deficit multiplied by the length.
}
