% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsec_utils.R
\name{rr_get_subseg_defects_breakdown}
\alias{rr_get_subseg_defects_breakdown}
\title{Gets a Breakdown of Defects contributing to Deficit Score}
\usage{
rr_get_subseg_defects_breakdown(sub_seg, deficit_data, benefit_scaler)
}
\arguments{
\item{sub_seg}{row from sub-sectioning output data frame. This data
frame should contain all distresses or data relating to capacity deficits,
and should contain the following columns:
\enumerate{
\item section_id: containing the unique ID for sections on which deficit
data is found
\item loc_from: start location for each deficit observation
\item loc_to: end location for each deficit observation
\item deficit: normalised deficit score for each observation. It is
recommended that a scale of 0 to 10 be used, with higher values indicating
greater deficit (distress or capacity deficit)
\item data_code: short code identifying each data type in the set
}}

\item{deficit_data}{data frame with deficit data for the network. This data
frame should contain all distresses or data relating to capacity deficits,
and should contain the following columns:
\enumerate{
\item section_id: containing the unique ID for sections on which deficit
data is found
\item loc_from: start location for each deficit observation
\item loc_to: end location for each deficit observation
\item deficit: normalised deficit score for each observation. It is
recommended that a scale of 0 to 10 be used, with higher values indicating
greater deficit (distress or capacity deficit)
\item data_code: short code identifying each data type in the set
}}

\item{benefit_scaler}{scaling factor for deficits. This factor should be
manually calibrated using the plotting function}
}
\description{
\code{rr_get_subseg_defects_breakdown} analyses the deficits within a
particular segment based on the 'data_code' for each deficit, and returns
a code that that represents the top contributing defects
}
