% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_utils.R
\name{rr_get_tl_data}
\alias{rr_get_tl_data}
\title{Gets all data for a specific treatment length}
\usage{
rr_get_tl_data(treat_len, dt)
}
\arguments{
\item{treat_len}{data frame containing information for the specified
treatment length. Required columns are:
\enumerate{
\item section_id: ID for the section on which the treatment length is
\item loc_from: Start/From location for the treatment length
\item loc_to: End/To location for the treatment length
\item lane: Lane code for the treatment length
}}

\item{dt}{data table with data to filter for the treatment length. This
data table should contain the following columns:
\enumerate{
\item section_id: containing the unique ID for sections on which deficit
data is found
\item loc_from: start location for each deficit observation
\item loc_to: end location for each deficit observation
\item lane: data lane code, or 'all' if the data applies to all lanes
}}
}
\description{
\code{rr_get_tl_data} Filters data and returns only those rows where
(a) the section Id matches the section ID of the specified treatment length;
(b) the data locations has some overlap with treatment length locations (note
that there is no minimum overlap specified)
(c) the data lane code is 'all' or else matches the lane code of the
specified treatment length (if the treatment length has lane code 'all', then
lane filtering is omitted)
}
