% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engineering.R
\name{rr_overlap_lengths}
\alias{rr_overlap_lengths}
\title{Gets the overlaps between one linear segment and a set of others}
\usage{
rr_overlap_lengths(seg1_start, seg1_end, seg_starts, seg_ends)
}
\arguments{
\item{seg1_start}{'start' or 'from' location for the first segment}

\item{seg1_end}{'end' or 'to' location for the first segment}

\item{seg_starts}{'start' or 'from' location for the segments to check
against the first segment}

\item{seg_ends}{'end' or 'to' location for the segments to check against the
first segment}
}
\value{
voctor containing lengths of the overlap with first section
}
\description{
\code{rr_overlap_lengths} is a vectorised version for getting the overlaps
between one segment and a set of others. Takes the start and end locations
for a linear segment plus arrays of the starts and ends of a set of other
segments and returns the length of the overlap between the first segment
and each of the other segments in the same units as the start and end
locations
}
