% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_visual_surv.R
\name{rr_spread_visual_distress}
\alias{rr_spread_visual_distress}
\title{Converts a visual distress percentage to approximated local distress areas}
\usage{
rr_spread_visual_distress(
  segments,
  defect_pct_column,
  defect_label,
  tl_id_col,
  section_id_column,
  area_name_column,
  lane_column,
  loc_from_col,
  loc_to_col,
  n_splits = 10,
  min_split_length = 20
)
}
\arguments{
\item{segments}{data frame with the treatment lengths and distress percentages}

\item{defect_pct_column}{name of the column containing the defect percentages}

\item{defect_label}{label to assign to observations, e.g. "viz_croc_crax"}

\item{tl_id_col}{name of the column containing the treatment length ID}

\item{section_id_column}{name of the column containing the section IDs}

\item{area_name_column}{name of the column containing the treatment length
area names}

\item{lane_column}{name of the column containing the lane codes}

\item{loc_from_col}{name of the column containing the start locations for each
treatment length}

\item{loc_to_col}{name of the column containing the end locations for each
treatment length}

\item{n_splits}{number of splits to do on each treatment length}

\item{min_split_length}{minimum length below which treatment length will not
be split (instead, length will be shortened to match distress area)}
}
\value{
vector of percentage area values, with percentages as 0 to 100+ values.
}
\description{
\code{rr_spread_visual_distress} takes a visual distress over a treatment
length and then splits it into approximated individual distress areas. This
is needed to incorporate traditional visual distress percentages to an
approximate set of local distress areas. This function is needed to utilise
the traditional visual assessments in the sub-sectioning tool which uses
individual distress locations. It is an approximate solution but better than
disregarding visual survey results completely.
}
\details{
For example, if a treatment length is from zero to 200m and has 20\% of a
specific distress, then this function will create \code{n_splits} sub-segments
each with a distress that covers 20\% of its length. So if \code{n_splits} is
10 (the default), then this function will split this treatment length-wide
distress into 10 "distresses", each one covering 20\% of the length. Thus,
first segment is 0 to 20 m (10th of total length), of which 20\% has distress,
thus distress assumed to be from 0 to 4. Second is from 20 to 40, thus distress
is assumed to be from 20 to 24, etc.

The output is a 'long' set containing simulated distress areas within each
treatment length, such that the total distress areas matches the original
percentage distress. The result set will not contain any observations for
treatment lengths where the distress percentage was zero.

Note that because of rounding of lengths and sub-lengths, and also because
any distress length less than 1 is droppped, the set of simulated distress
areas will typically be shorter than the total length of areas based on the
percentage of distress on total treatment lengths. Again, this is an
approximate method to simulate individual distresses based on estimated total
areas of distress.
}
