% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsec_utils.R
\name{rr_subsecs_fill_gaps}
\alias{rr_subsecs_fill_gaps}
\title{Fills any gaps between sub-sections to cover entire length}
\usage{
rr_subsecs_fill_gaps(sub_segs, sec_start, sec_end)
}
\arguments{
\item{sub_segs}{data frame with identified sub-sections and scores. This
should be the output from method \code{tt_get_subsections_v1}}

\item{sec_start}{start location for the section or treatment length}

\item{sec_end}{end location for the section or treatment length}
}
\value{
data frame containing the original sub-sections but with added rows
to ensure all gaps between sub-sections and start and end for the treatment
length or section are covered
}
\description{
\code{rr_subsecs_fill_gaps} fills any gaps between sub-sections, or between
the start of a section or treatment length and start of the first
sub-section, or between the end of the section or treatment length and the
end of the last sub-section.
}
\details{
Rows are added to fill the gaps and the score and index values for these
'filler' sub-segments are set to -1.
}
