% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsec_utils.R
\name{rr_subsecs_join_ends}
\alias{rr_subsecs_join_ends}
\title{Checks sub-sections and joins ends to start and end of section if close}
\usage{
rr_subsecs_join_ends(sub_segs, sec_start, sec_end, max_gap = 150)
}
\arguments{
\item{sub_segs}{data frame with identified sub-sections and scores. This
should be the output from method \code{tt_get_subsections_v1}}

\item{sec_start}{start location for the section}

\item{sec_end}{end location for the section}

\item{max_gap}{maximum gap between the start or end of a sub-section and the
start/end of the section. Default is 150 m.}
}
\value{
data frame containing the original sub-sections but with adjusted
start and ends for the first and last sub-sections if these are within
\code{max_gap} of the section start or end
}
\description{
\code{rr_subsecs_join_ends} Checks the set of sub-sections and joins the
first and last sub-section limits to the start and/or end of the section if
these are within range.
}
\details{
This function can be used to modify sub-sections by ensuring there are not
short sections between a sub-section to be treated and the start or end of
a section
}
