% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_joiner1.R
\name{rr_subsecs_trim_shorts}
\alias{rr_subsecs_trim_shorts}
\title{Trims any short segments by joining it to adjacent segments}
\usage{
rr_subsecs_trim_shorts(segments, min_length, value_column)
}
\arguments{
\item{segments}{segments from which to remove/join too short segments}

\item{min_length}{minimum allowed length}

\item{value_column}{name of the column that contains the score or value that
will determine which neighbour is 'most-alike'. The value in this column will
be updated based on a weighted sum of the lengths before joining.}
}
\value{
data frame containing the segments will locations and values
adjusted so that there are no segments with length below the minimum allowed
length
}
\description{
\code{rr_subsecs_trim_shorts} trims out any segments shorter than a
specified length by joining it to one of it's two neighbours. To determine
which of the two neighbours are most appropriate to join to, it compares the
value in the column '\code{value_column}' and picks the neighbour with the
value closest to the current value. A
}
\details{
After joining the segment to the most appropriate neighbour, the value in the
value column is updating using a length based weighted sum. For example, if
the short segment being joined has a length of 50 and a value of 10, and
the selected neighbour has a length of 200 and a value of 12, the joined
segment will get a value of \code{(50/250)*10 + (200/250)*12 = 11.6}
}
