% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_sectioning_v1.R
\name{tt_plot_segment_data}
\alias{tt_plot_segment_data}
\title{Plots the data and sub-section locations for a section}
\usage{
tt_plot_segment_data(
  data,
  sub_segs,
  x_min = NULL,
  x_max = NULL,
  y_min = 0,
  y_max = 12,
  best_seg_y = 10.5
)
}
\arguments{
\item{data}{data frame with deficit data for the section}

\item{sub_segs}{data frame with identified sub-sections and scores. This
should be the output from method \code{tt_get_subsections_v1}}

\item{x_min}{minimum for x-axis scale}

\item{x_max}{maximum for x-axis scale}

\item{y_min}{minimim for y-axis scale. Default is zero, assuming a scale from
1 to 12}

\item{y_max}{maximim for y-axis scale. Default is 12, assuming a scale from
1 to 10 for deficits and then allows a bit at the top}

\item{best_seg_y}{y-location for the top-scoring sub-section. Default is 10.5
for a scale from 0 to 12}
}
\value{
ggplot object that can be further customised
}
\description{
\code{tt_plot_segment_data} Plots the data and sub-section locations (with
their scores) for a section
}
